requirements.txt
arch==5.3.1
image_quality==1.2.7
lpips==0.1.4
matplotlib==3.5.3
numpy==1.18.5
opencv_contrib_python==4.7.0.72
Pillow==9.5.0
scipy==1.7.3
thop==0.1.1.post2209072238
torch==1.13.1+cu117
torchvision==0.14.1
tqdm==4.65.0
einops==0.6.1